Final fixes

+++ scripts/linux.sh
@@ -9,19 +9,20 @@ else
 fi
 
 echo "Installing udev configuration files..."
-dart run :udev
+dart run :udev >&3
 sudo cp linux/*.rules /etc/udev/rules.d
 sudo udevadm control --reload-rules
 sudo udevadm trigger
 
 echo "Installing systemd configuration files..."
-dart run :systemd  # generate the service files
+dart run :systemd  >&3
 for service in $(find linux -name "*.service" -print0 | xargs -0 basename -a -s .service)
 do
-  # Stop the service, overwrite it, re-enable it
-  sudo systemctl stop $service    >&3
-  sudo systemctl disable $service >&3
-  sudo cp $service.service /etc/systemd/system >&3
+  if [[ $(sudo systemctl is-active $service) == "active" ]]; then
+    sudo systemctl stop $service    >&3
+    sudo systemctl disable $service >&3
+  fi
+  sudo cp linux/$service.service /etc/systemd/system >&3
   sudo systemctl enable $service  >&3
   sudo systemctl start $service   >&3
 done
